refactor(chain): RPC-failover read-intent + stickiness ergonomics (#1548)
Behavior-preserving follow-up to #1544. No production behavior change; all
packages/chain unit tests (600 pass / 1 skip), mock-adapter parity, and tsc
pass unchanged — the only test diff is relocating the Mechanism-B AC suite.

- hub-rotation-poller: fold the 3 background tip probes behind a `readTip`
  wrapper so no call site spells the transport-internal `skipPreferred` flag.
- endpoint-stickiness: add `attempts(canonical, intent)` bundling the per-loop
  order + recordSuccess + recordFailure triple; rewire all 4 failover loops so
  (canonical, intent) can't drift across the three calls. order() still runs
  exactly once per loop entry.
- evm-adapter-conviction: collapse the PCA requestPublishingConvictionRpc
  4-branch method/params ladder into a pure classifyPcaRead classifier + one
  dispatch (identical per-method routing).
- test: split the ~580-line endpoint-stickiness AC suite (AC-1..AC-29) out of
  rpc-failover-client.unit.test.ts into a dedicated acceptance file.
