+++ packages/chain/src/rpc-failover-client.ts
@@ -369,7 +369,8 @@ export class RpcFailoverClient {
     // 'nonceWrite': the state machine starts a fresh-nonce populate on a backend
     // ONLY if a prior populate proved it (else canonical primary-first = the
     // authoritative nonce source) — the nonce-staleness guard.
-    const endpoints = this.stickiness.order(canonical, 'nonceWrite');
+    const plan = this.stickiness.attempts(canonical, 'nonceWrite');
+    const endpoints = plan.ordered;
     let lastRetryable: unknown;
     for (let i = 0; i < endpoints.length; i += 1) {
       const endpoint = endpoints[i];
@@ -420,12 +421,12 @@ export class RpcFailoverClient {
         // Signed on this endpoint → 'nonceWrite' marks it WRITE-proven (nonce-safe)
         // so it's preferred for the read-your-write ops that follow (the caller's
         // broadcast + receipt + confirming re-reads) AND for subsequent populates.
-        this.stickiness.recordSuccess(endpoint, canonical, 'nonceWrite', i === 0);
+        plan.recordSuccess(endpoint, i);
         return signed;
       } catch (err) {
         if (!isRetryableRpcError(err)) throw err;
         lastRetryable = err;
-        this.stickiness.recordFailure(endpoint, 'nonceWrite'); // de-prefer a failed backend
+        plan.recordFailure(endpoint); // de-prefer a failed backend
         if (i < endpoints.length - 1) {
           noteRpcFailover(`${label} preparation`, endpoints[i].rpcUrl, err, endpoints[i + 1].rpcUrl);
         }
@@ -472,7 +473,8 @@ export class RpcFailoverClient {
         const startedAt = Date.now();
         try {
           const canonical = this.getEndpoints();
-          const endpoints = this.stickiness.order(canonical, 'write');
+          const plan = this.stickiness.attempts(canonical, 'write');
+          const endpoints = plan.ordered;
           let lastRetryable: unknown;
           for (let i = 0; i < endpoints.length; i += 1) {
             const endpoint = endpoints[i];
@@ -491,7 +493,7 @@ export class RpcFailoverClient {
               );
               span.setAttribute('dkg.tx_hash', txHash);
               this.recordRpcOutcome('eth_sendRawTransaction', 'ok');
-              this.stickiness.recordSuccess(endpoint, canonical, 'write', i === 0);
+              plan.recordSuccess(endpoint, i);
               return;
             } catch (err) {
               if (isKnownTransactionError(err)) {
@@ -499,7 +501,7 @@ export class RpcFailoverClient {
                 span.setAttribute('dkg.tx_hash', txHash);
                 span.addEvent('broadcast.already_known', { attempt: i + 1 });
                 this.recordRpcOutcome('eth_sendRawTransaction', 'ok');
-                this.stickiness.recordSuccess(endpoint, canonical, 'write', i === 0);
+                plan.recordSuccess(endpoint, i);
                 return;
               }
               if (!isRetryableRpcError(err)) {
@@ -507,7 +509,7 @@ export class RpcFailoverClient {
                 throw err;
               }
               lastRetryable = err;
-              this.stickiness.recordFailure(endpoint, 'write'); // de-prefer a failed backend
+              plan.recordFailure(endpoint); // de-prefer a failed backend
               if (i < endpoints.length - 1) {
                 noteRpcFailover(`${label} broadcast`, endpoints[i].rpcUrl, err, endpoints[i + 1].rpcUrl);
               }
@@ -553,7 +555,8 @@ export class RpcFailoverClient {
         const startedAt = Date.now();
         try {
           const canonical = this.getEndpoints();
-          const endpoints = this.stickiness.order(canonical, 'write');
+          const plan = this.stickiness.attempts(canonical, 'write');
+          const endpoints = plan.ordered;
           let lastRetryable: unknown;
           let sawNonErrorResponse = false;
           for (let i = 0; i < endpoints.length; i += 1) {
@@ -579,7 +582,7 @@ export class RpcFailoverClient {
                 // it. A null "not mined yet" response is NOT a stickiness signal
                 // (no single winning endpoint), so we only note on a real
                 // receipt — the self-heal still polls every endpoint per tick.
-                this.stickiness.recordSuccess(endpoint, canonical, 'write', i === 0);
+                plan.recordSuccess(endpoint, i);
                 return receipt;
               }
             } catch (err) {
@@ -588,7 +591,7 @@ export class RpcFailoverClient {
                 throw err;
               }
               lastRetryable = err;
-              this.stickiness.recordFailure(endpoint, 'write'); // de-prefer a failed backend
+              plan.recordFailure(endpoint); // de-prefer a failed backend
               if (i < endpoints.length - 1) {
                 noteRpcFailover('receipt lookup', endpoints[i].rpcUrl, err, endpoints[i + 1].rpcUrl);
               }
@@ -647,7 +650,8 @@ export class RpcFailoverClient {
     // A tip-sensitive read (skipPreferred) is 'transparentRead' (canonical order +
     // no state mutation); a normal read is 'stickyRead'.
     const intent: StickinessIntent = skipPreferred ? 'transparentRead' : 'stickyRead';
-    const endpoints = this.stickiness.order(canonical, intent);
+    const plan = this.stickiness.attempts(canonical, intent);
+    const endpoints = plan.ordered;
     const capMs = resolveCapMs(policy, canonical.length);
     let lastRetryable: unknown;
     let sawEmpty = false;
@@ -677,12 +681,12 @@ export class RpcFailoverClient {
           lastEmpty = out;
           continue;
         }
-        this.stickiness.recordSuccess(endpoint, canonical, intent, i === 0);
+        plan.recordSuccess(endpoint, i);
         return out;
       } catch (err) {
         if (!isRetryable(err)) throw err;
         lastRetryable = err;
-        this.stickiness.recordFailure(endpoint, intent); // de-prefer a failed backend
+        plan.recordFailure(endpoint); // de-prefer a failed backend
         if (!isLast) {
           noteRpcFailover(label, endpoints[i].rpcUrl, err, endpoints[i + 1].rpcUrl);
         }
